moved everything to float

+++ src/Core.cpp
@@ -103,7 +103,7 @@ void Core::save()
         for(unsigned int j = 0; j < nodes[i]->output.size(); j++)
         {
             saveXml.addTag("PIN");
-            saveXml.addAttribute("PIN", "VALUE", boost::get<double>(nodes[i]->output[j]->value), j);
+            saveXml.addAttribute("PIN", "VALUE", boost::get<float>(nodes[i]->output[j]->value), j);
         }
         saveXml.popTag();
     }
@@ -132,7 +132,7 @@ void Core::load()
         loadXml.pushTag("NODE", i);
         for(int j = 0; j < loadXml.getNumTags("PIN"); j++)
         {
-            temp->output[j]->value = (double)ofToFloat(loadXml.getAttribute("PIN","VALUE", "", j).c_str());
+            temp->output[j]->value = ofToFloat(loadXml.getAttribute("PIN","VALUE", "", j).c_str());
             temp->init();
         }
         loadXml.popTag();
