+++ src/Slider.h
@@ -6,7 +6,7 @@
 class Slider
 {
     public:
-        Slider(float* _x, float* _y, string _name, double* _val, bool* _bIsNodeActive);
+        Slider(float* _x, float* _y, string _name, float* _val, bool* _bIsNodeActive);
         ~Slider();
         void draw(ofEventArgs & args);
         void setValue(float mx, float my, bool bCheck);
@@ -16,7 +16,7 @@ class Slider
         virtual void mouseDragged(ofMouseEventArgs & args);
         virtual void mouseReleased(ofMouseEventArgs & args);
 
-        double* val;
+        float* val;
         float* x, * y;
         float min, max;
         string name;
