+++ src/nodes/Multiply.cpp
@@ -18,6 +18,6 @@ Multiply::~Multiply()
 void Multiply::process()
 {
     //cout << "process from Multiply()\n";
-    output[0]->value = boost::get<double>(input[0]->value) * boost::get<double>(input[1]->value);
+    output[0]->value = boost::get<float>(input[0]->value) * boost::get<float>(input[1]->value);
     //cout << "output pin: " << output[0]->value << endl;
 }
